refactor: 提取出getDirtyFile到ContextAnalysisDiagram
